requirements.txt
numpy==1.19.2
pillow==8.0.1